sem10/3D-proj/requirements.txt
numpy~=1.24
scipy~=1.10.1
Pillow~=9.4
PyOpenGL~=3.1
PyOpenGL-accelerate~=3.1
glfw~=2.5
cython~=0.29
AssimpCy~=2.1
